BMC add hosts file restructure
