+++ libs/ui-lib/lib/cim/components/Agent/index.ts
@@ -1,3 +1,3 @@
 export { default as AgentTable } from './AgentTable';
-export { default as BMCForm } from './BMCForm';
+export { default as BMCForm } from './BMCForm/BMCForm';
 export * from './types';
